Register new data access classes for dependency injection and restructure interfaces into MixMate.Core project
