+++ MixMate.Core/Interfaces/ISongRepository.cs
@@ -1,0 +1,9 @@
+﻿using MixMate.Core.Entities;
+
+namespace MixMate.Core.Interfaces;
+
+public interface ISongRepository
+{
+    Task AddSongsAsync(List<Song> songs);
+    Task<IEnumerable<Song>> GetAllSongsAsync();
+}
