+++ MixMate.Web/MixMate.Web.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\MixMate.Core\MixMate.Core.csproj" />
+    <ProjectReference Include="..\MixMate.DataAccess\MixMate.DataAccess.csproj" />
   </ItemGroup>
 
 </Project>
