+++ MixMate.Web/Program.cs
@@ -1,4 +1,7 @@
+using MixMate.Core.Interfaces;
 using MixMate.Core.Services;
+using MixMate.DataAccess.Database;
+using MixMate.DataAccess.Repositories;
 using MixMate.Web.Components;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -9,6 +12,16 @@ builder.Services.AddRazorComponents()
 
 // Register services
 builder.Services.AddScoped<IFileProcessingService, FileProcessingService>();
+
+// Register IDatabaseContext
+var connectionString = builder.Configuration.GetConnectionString("DefaultConnection")
+    ?? throw new InvalidOperationException("The DefaultConnection connection string is not configured.");
+
+builder.Services.AddSingleton<IDatabaseContext>(provider =>
+    new DatabaseContext(connectionString));
+
+// Register repositories
+builder.Services.AddScoped<ISongRepository, SongRepository>();
 
 var app = builder.Build();
